requirements.txt
google-generativeai
python-dotenv
streamlit
pyperclip
Pillow
google-adk